# bloom: reduce memory allocations in table bloom filter