#### bloom: add a table bloom filter benchmark


#### bloom: reduce memory allocations in table bloom filter

A minor performance improvement to avoid many allocations of the
`hashes` buffer in `tableFilterWriter`.

We extend `FilterPolicy` with a `DropWriter` method which allows reuse
of writers. The implementation uses a `sync.Pool`.

```
name            old time/op    new time/op    delta
BloomFilter-16    45.4µs ± 0%    43.8µs ± 1%   -3.50%  (p=0.000 n=10+10)

name            old alloc/op   new alloc/op   delta
BloomFilter-16    10.0kB ± 0%     1.8kB ± 0%  -82.03%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
BloomFilter-16      11.0 ± 0%       1.0 ± 0%  -90.91%  (p=0.000 n=10+10)
```

Fixes #1999